app/requirements.txt
Django==2.2.10
djangorestframework==3.9.2
psycopg2==2.7.3.2
gunicorn==19.7.1
gevent==1.4.0
requests==2.23.0